+++ check_cores.cpp
@@ -15,7 +15,7 @@ int main() {
 
     string graph6;
     clock_t c_start = clock();
-    int counter = 0;
+    long counter = 0;
 
     while(getline(cin, graph6)) {
         if (!filter(graph6)) {
